fixed Point3 typedef
git-svn-id: https://scm.citationtech.net/trunk/shared/math/math@567 b7e6147c-0f76-4c84-a09d-19069afaf10c

+++ math/geometry_core.hpp
@@ -80,7 +80,7 @@ typedef Point2f Point2;
 
 typedef Point3_<int> Point3i;
 typedef Point3_<double> Point3f;
-typedef Point3i Point3;
+typedef Point3f Point3;
 
 typedef std::vector<Point2_<int> > Points2i;
 typedef std::vector<Point2_<double> > Points2f;
